chore: refactor the styling of selection and hovering in TreeMode (WIP)
