+++ src/lib/components/controls/contextmenu/ContextMenuPointer.scss
@@ -13,12 +13,20 @@
   background: transparent;
 
   border-radius: 2px;
-  background: $context-menu-pointer-background;
+  background: $context-menu-pointer-hover-background;
   color: $context-menu-pointer-color;
   border: none;
   box-shadow: $controls-box-shadow;
 
   &:hover {
     background: $context-menu-pointer-background-highlight;
+  }
+
+  &.jse-selected {
+    background: $context-menu-pointer-background;
+
+    &:hover {
+      background: $context-menu-pointer-background-highlight;
+    }
   }
 }
