+++ src/lib/components/modes/tablemode/JSONValueCell.svelte
