+++ src/lib/components/modes/tablemode/TableMode.scss
@@ -114,8 +114,17 @@
             // FIXME: border and/or border shadow?
           }
 
-          &.jse-selected-value > :global(.jse-value) {
-            background: $selection-background-color;
+          .jse-value-outer {
+            display: inline-block;
+            cursor: $contents-cursor;
+
+            &:hover {
+              background: $hover-background-color;
+            }
+
+            &.jse-selected-value {
+              background: $selection-background-color;
+            }
           }
 
           // FIXME: this is ugly. give JSONValue an extra class instead?
