+++ src/lib/components/modes/tablemode/TableMode.svelte
@@ -70,7 +70,7 @@
     toTableCellPosition
   } from '$lib/logic/table.js'
   import { isEmpty, isEqual, uniqueId } from 'lodash-es'
-  import JSONValueComponent from './JSONValue.svelte'
+  import JSONValueCell from './JSONValueCell.svelte'
   import {
     activeElementIsChildOf,
     createNormalizationFunctions,
@@ -1790,49 +1790,52 @@
                     isValueSelection(selection) && pathStartsWith(selection.path, path)}
                   {@const validationErrorsByColumn = validationErrorsByRow?.columns[columnIndex]}
                   {@const validationError = mergeValidationErrors(path, validationErrorsByColumn)}
-                  <td
-                    class="jse-table-cell"
-                    data-path={encodeDataPath(path)}
-                    class:jse-selected-value={isSelected}
-                  >
-                    {#if isObjectOrArray(value)}
-                      {@const searchResultsByCell = flattenSearchResults(
-                        getInRecursiveState(item, searchResultByRow, column)
-                      )}
+                  <td class="jse-table-cell" data-path={encodeDataPath(path)}>
+                    <div class="jse-value-outer" class:jse-selected-value={isSelected}>
+                      {#if isObjectOrArray(value)}
+                        {@const searchResultsByCell = flattenSearchResults(
+                          getInRecursiveState(item, searchResultByRow, column)
+                        )}
 
-                      {@const containsActiveSearchResult = searchResultsByCell
-                        ? searchResultsByCell.some((item) => item.active)
-                        : false}
+                        {@const containsActiveSearchResult = searchResultsByCell
+                          ? searchResultsByCell.some((item) => item.active)
+                          : false}
 
-                      <InlineValue
-                        {path}
-                        {value}
-                        {parser}
-                        {isSelected}
-                        containsSearchResult={!isEmpty(searchResultsByCell)}
-                        {containsActiveSearchResult}
-                        onEdit={openJSONEditorModal}
-                      />{:else}
-                      {@const searchResultItemsByCell = getInRecursiveState(
-                        json,
-                        searchResults,
-                        path
-                      )?.searchResults}
+                        <InlineValue
+                          {path}
+                          {value}
+                          {parser}
+                          {isSelected}
+                          containsSearchResult={!isEmpty(searchResultsByCell)}
+                          {containsActiveSearchResult}
+                          onEdit={openJSONEditorModal}
+                        />{:else}
+                        {@const searchResultItemsByCell = getInRecursiveState(
+                          json,
+                          searchResults,
+                          path
+                        )?.searchResults}
 
-                      <JSONValueComponent
-                        {path}
-                        value={value !== undefined ? value : ''}
-                        enforceString={getEnforceString(json, documentState, path, context.parser)}
-                        selection={isSelected ? selection : undefined}
-                        searchResultItems={searchResultItemsByCell}
-                        {context}
-                      />{/if}{#if !readOnly && isSelected && !isEditingSelection(selection)}
-                      <div class="jse-context-menu-anchor">
-                        <ContextMenuPointer selected={true} onContextMenu={openContextMenu} />
-                      </div>
-                    {/if}{#if validationError}
-                      <ValidationErrorIcon {validationError} onExpand={noop} />
-                    {/if}
+                        <JSONValueCell
+                          {path}
+                          value={value !== undefined ? value : ''}
+                          enforceString={getEnforceString(
+                            json,
+                            documentState,
+                            path,
+                            context.parser
+                          )}
+                          selection={isSelected ? selection : undefined}
+                          searchResultItems={searchResultItemsByCell}
+                          {context}
+                        />{/if}{#if !readOnly && isSelected && !isEditingSelection(selection)}
+                        <div class="jse-context-menu-anchor">
+                          <ContextMenuPointer selected={true} onContextMenu={openContextMenu} />
+                        </div>
+                      {/if}{#if validationError}
+                        <ValidationErrorIcon {validationError} onExpand={noop} />
+                      {/if}
+                    </div>
                   </td>
                 {/each}
                 {#if showRefreshButton}
