+++ src/lib/components/modes/treemode/JSONNode.svelte
@@ -573,7 +573,6 @@
   style:--level={path.length}
   class:jse-root={root}
   class:jse-selected={isNodeSelected && isMultiSelection(selection)}
-  class:jse-selected-key={isNodeSelected && isKeySelection(selection)}
   class:jse-selected-value={isNodeSelected && isValueSelection(selection)}
   class:jse-readonly={context.readOnly}
   class:jse-hovered={hover === HOVER_COLLECTION}
@@ -806,10 +805,12 @@
             ? validationErrors.properties[key]
             : undefined}
 
+          {@const nestedPath = path.concat(key)}
+
           {@const nestedSelection = selectionIfOverlapping(
             context.getJson(),
             selection,
-            path.concat(key)
+            nestedPath
           )}
 
           <svelte:self
@@ -822,7 +823,12 @@
             {context}
             onDragSelectionStart={handleDragSelectionStart}
           >
-            <div slot="identifier" class="jse-identifier">
+            <div
+              slot="identifier"
+              class="jse-key-outer"
+              class:jse-selected-key={isKeySelection(nestedSelection) &&
+                isEqual(nestedSelection.path, nestedPath)}
+            >
               <JSONKey
                 pointer={propPointer}
                 {key}
@@ -856,14 +862,16 @@
         {#if !root}
           <div class="jse-separator">:</div>
         {/if}
-        <JSONValue
-          {path}
-          {value}
-          {enforceString}
-          selection={isNodeSelected ? selection : undefined}
-          searchResultItems={filterValueSearchResults(searchResults)}
-          {context}
-        />
+        <div class="jse-value-outer">
+          <JSONValue
+            {path}
+            {value}
+            {enforceString}
+            selection={isNodeSelected ? selection : undefined}
+            searchResultItems={filterValueSearchResults(searchResults)}
+            {context}
+          />
+        </div>
         {#if !context.readOnly && isNodeSelected && selection && (isValueSelection(selection) || isMultiSelection(selection)) && !isEditingSelection(selection) && isEqual(getFocusPath(selection), path)}
           <div class="jse-context-menu-pointer-anchor">
             <ContextMenuPointer selected={true} onContextMenu={context.onContextMenu} />
